+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 
+## 0.9.5
+
+- small rework on database
+- API: creation of new share items
+
+
 ## 0.9.4
 
 - Fixed an SQL error (#51)